+++ .eslintrc
@@ -27,6 +27,7 @@
     "import/extensions": "off",
     "no-undef": "warn",
     "react/jsx-props-no-spreading": "off",
-    "react/require-default-props": "off"
+    "react/require-default-props": "off",
+    "@typescript-eslint/no-use-before-define": "off"
   }
 }
